raspberry_pi/requirements.txt
pyserial==3.5
RPi.GPIO==0.7.1
websockets==12.0
